+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/momo/service/SurePayMomoPaymentIntegrationWritePlatformService.java
@@ -16,10 +16,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.fineract.portfolio.loanaccount.domain;
+package org.apache.fineract.infrastructure.momo.service;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import java.io.IOException;
+import org.apache.fineract.portfolio.loanaccount.data.MomoPaymentData;
+import org.apache.fineract.portfolio.loanaccount.domain.Loan;
+import org.apache.fineract.portfolio.loanaccount.domain.LoanTransaction;
 
-public interface MomoLoanPaymentTransactionRepository
-        extends JpaRepository<MomoLoanPaymentTransaction, Long>, JpaSpecificationExecutor<MomoLoanPaymentTransaction> {}
+public interface SurePayMomoPaymentIntegrationWritePlatformService {
+
+    void payOut(MomoPaymentData momoPaymentData, Loan loan, LoanTransaction loanTransaction) throws IOException;
+
+    void processLoanTransactionsOnMomoPayment();
+}
